Moved insurance to user folder
