+++ src/App.jsx
@@ -20,12 +20,12 @@ import SearchResults from './components/pages/SearchResults'
 import Main from './components/pages/user/main/Main'
 import Prescription from './components/pages/Prescription'
 import SignUp from './components/pages/signUp/SignUp'
-import InsuranceCompanies from './components/pages/insurance/InsuranceCompanies'
-import CompanyPage from './components/pages/insurance/CompanyPage'
-import InsuranceCardPage from './components/pages/insurance/InsuranceCardPage'
+import InsuranceCompanies from './components/pages/user/insurance/InsuranceCompanies'
+import CompanyPage from './components/pages/user/insurance/CompanyPage'
+import InsuranceCardPage from './components/pages/user/insurance/InsuranceCardPage'
 import NewPassword from './components/pages/login/NewPassword'
 import Map from './components/pages/user/map/Map'
-import AddCard from './components/pages/insurance/AddCard'
+import AddCard from './components/pages/user/insurance/AddCard'
 import CategoryPage from './components/pages/categoryPage/CategoryPage'
 import ErrorPage from './components/common/ErrorPage'
 import Pharmacy from './components/pages/pharmacy/Pharmacy'
